requirements.txt
sqlalchemy
aiosqlite

fastapi
uvicorn

pydantic
pydantic-settings

pyjwt[crypto]
bcrypt